requirements_local.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
colorama==0.4.6
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
